Move helper functions to test_utils and remove static PARAMETERS
